Use _parse_array_or_* whenever appropriate
This patch fixes some issues when the argument of an operator is a
literal array with reference to some variable or document path (e.g.
['$foo.bar', 1]).

Note that not all places where we made the replacement check whether the
returned type is really a list, so this patch also addd a TODO comment
so we do not miss that.

This commit should to fix #700 ("$concatArrays doesn't seem to handle
when it should replace variable inside a list"). A test case is added
for such scenario. We also add a similar test case for the $map
operator.